# Pipeline parallel idle optimization