--- a/python/sglang/srt/server_args.py
+++ b/python/sglang/srt/server_args.py
@@ -588,6 +588,7 @@ class ServerArgs:
     enable_mscclpp: bool = False
     enable_torch_symm_mem: bool = False
     disable_overlap_schedule: bool = False
+    enable_pp_sleep_on_idle: bool = False
     enable_mixed_chunk: bool = False
     enable_dp_attention: bool = False
     enable_dp_lm_head: bool = False
@@ -2192,6 +2193,14 @@ class ServerArgs:
             self.disable_overlap_schedule = True
             logger.warning(
                 "Pipeline parallelism is incompatible with overlap schedule."
+            )
+
+        if self.enable_pp_sleep_on_idle and (
+            (pp_size_per_node := max(self.pp_size // self.nnodes, 1)) <= 1
+        ):
+            self.enable_pp_sleep_on_idle = False
+            logger.warning(
+                "Pipeline parallelism sleep on idle is only compatible for pipeline parallelism occurring within nodes."
             )
 
     def _handle_hicache(self):
@@ -4436,6 +4445,11 @@ class ServerArgs:
             help="Disable the overlap scheduler, which overlaps the CPU scheduler with GPU model worker.",
         )
         parser.add_argument(
+            "--enable-pp-sleep-on-idle",
+            action="store_true",
+            help="Enable pipeline parallel sleep while idle, preventing costly GPU polling during idle periods.",
+        )
+        parser.add_argument(
             "--enable-mixed-chunk",
             action="store_true",
             help="Enabling mixing prefill and decode in a batch when using chunked prefill.",
@@ -5495,6 +5509,15 @@ class PortArgs:
     # The ipc filename for Tokenizer and worker tokenizer
     tokenizer_worker_ipc_name: Optional[str]
 
+    # The ipc filenames for PP ranks succeeding other stages on same node (zmq)
+    pp_idle_wakeup_listeners_ipc_names: Optional[
+        dict[tuple[Optional[int], int], str]
+    ] = dataclasses.field(default_factory=dict)
+    # The ipc filenames for PP ranks preceding other stages on same node (zmq)
+    pp_idle_wakeup_notifiers_ipc_names: Optional[
+        dict[tuple[Optional[int], int], str]
+    ] = dataclasses.field(default_factory=dict)
+
     @staticmethod
     def init_new(
         server_args: ServerArgs,
@@ -5530,6 +5553,8 @@ class PortArgs:
                 rpc_ipc_name=f"ipc://{tempfile.NamedTemporaryFile(delete=False).name}",
                 metrics_ipc_name=f"ipc://{tempfile.NamedTemporaryFile(delete=False).name}",
                 tokenizer_worker_ipc_name=tokenizer_worker_ipc_name,
+                pp_idle_wakeup_listeners_ipc_names={},
+                pp_idle_wakeup_notifiers_ipc_names={},
             )
         else:
             # DP attention. Use TCP + port to handle both single-node and multi-node.
@@ -5584,6 +5609,8 @@ class PortArgs:
                 rpc_ipc_name=f"tcp://{dist_init_host}:{rpc_port}",
                 metrics_ipc_name=f"tcp://{dist_init_host}:{metrics_ipc_name}",
                 tokenizer_worker_ipc_name=tokenizer_worker_ipc_name,
+                pp_idle_wakeup_listeners_ipc_names={},
+                pp_idle_wakeup_notifiers_ipc_names={},
             )