<!-- Thank you for your contribution! Please follow these guidelines to enhance your pull request. If anything is unclear, submit your PR and reach out to maintainers for assistance. Join our Slack community at https://slack.sglang.ai to discuss further. -->

## Motivation

Pipeline parallelism is causing non-primary GPU ranks to continuously busy-wait at near 100% utilization even while idle, consuming significant power unnecessarily. For example, on a 2x3090 single-node system with `--pp-size 2` PP rank 1 maintains ~350W power draw while idle due to NCCL's `cudaStreamSynchronize` polling behavior.

This PR enables PP stages to enter low-power sleep states when idle, reducing power consumption from ~350W to ~30W (>90% reduction) without impacting inference performance. This also reduces device heat, reducing throttling, which thereby has potential to provide performance gains.

Without this change, my device becomes useless for other tasks even when less than half of it is actually allocated for inference.

Please advise if I can provide any more detail and/or if any revisions are needed.

CCing pipeline parallelism owner @Ying1123 for review as per contribution guide.
  
<!-- Describe the purpose and goals of this pull request. -->

## Modifications

  - **Added `--enable-pp-sleep-on-idle` flag**: Opt-in feature for when `--pp-size > 1` and `pp_size_per_node > 1`
  - **Implemented CPU-only wake signaling**: Uses ZMQ PUSH/PULL sockets to coordinate PP stages without GPU involvement
    - PP rank N sends wake signal before sending work to PP rank N+1
    - PP rank N+1 waits on ZMQ socket instead of NCCL polling when idle
  - **Modified files**:
    - `scheduler.py`: Added wake socket initialization and idle wait logic
    - `server_args.py`: Added new flag and IPC mappings for wake sockets
    - `engine.py` & `data_parallel_controller.py`: Setup wake socket connections for same-node PP ranks
    - `docs/advanced_features/server_arguments.md`: Documented new flag
    - `test/srt/test_pp_single_node.py`: Added basic test
<!-- Detail the changes made in this pull request. -->

## Accuracy Tests

N/A
<!-- If this pull request affects model outputs (e.g., changes to the kernel or model forward code), provide accuracy test results. -->

## Benchmarking and Profiling

I compared `--enable-pp-sleep-on-idle` enabled (1) vs control from before any of my changes (2).

This test uses Qwen3-32B and benchmarks with following:
``` bash
docker compose --env-file .env -f docker-compose.sglang.yml run --rm \
    --entrypoint "python -m sglang.bench_serving" sglang \
    --backend sglang-oai \
    --host sglang \
    --port 8080 \
    --dataset-name random \
    --num-prompts 256 \
    --random-input-len 1024 \
    --random-output-len 512 \
    --random-range-ratio 0.2 \
    --max-concurrency 16 \
    --request-rate 16 \
    --warmup-requests 16 \
    --apply-chat-template \
    --extra-request-body '{"temperature":0.7,"top_p":0.8,"top_k":20,"min_p":0,"chat_template_kwargs":{"enable_thinking":false}}'
```

<img width="2061" height="956" alt="pr-comparison" src="https://github.com/user-attachments/assets/c5471e7f-80b3-4aed-aa68-d27109858558" />

Here, observe the period before and after the benchmark runs. For GPU 1 (`pp_rank > 1`) note the distinct and unrelenting spikes of power and utilization on the control. Note how with my flag enabled, the usage and power drops to true idling state.

### `--enable-pp-sleep-on-idle` enabled
```
  ============ Serving Benchmark Result ============
  Backend:                                 sglang-oai
  Traffic request rate:                    16.0
  Max request concurrency:                 16
  Successful requests:                     256
  Benchmark duration (s):                  609.87
  Total input tokens:                      155646
  Total generated tokens:                  80685
  Total generated tokens (retokenized):    80671
  Request throughput (req/s):              0.42
  Input token throughput (tok/s):          255.21
  Output token throughput (tok/s):         132.30
  Total token throughput (tok/s):          387.51
  Concurrency:                             15.53
  ----------------End-to-End Latency----------------
  Mean E2E Latency (ms):                   36987.83
  Median E2E Latency (ms):                 36250.98
  ---------------Time to First Token----------------
  Mean TTFT (ms):                          1459.29
  Median TTFT (ms):                        1152.60
  P99 TTFT (ms):                           6469.77
  ---------------Inter-Token Latency----------------
  Mean ITL (ms):                           113.09
  Median ITL (ms):                         71.93
  P95 ITL (ms):                            214.90
  P99 ITL (ms):                            1328.65
  Max ITL (ms):                            5737.65
  ==================================================
```

### Control (prior to my implementation)
```
============ Serving Benchmark Result ============
Backend:                                 sglang-oai
Traffic request rate:                    16.0
Max request concurrency:                 16
Successful requests:                     256
Benchmark duration (s):                  596.03
Total input tokens:                      155646
Total generated tokens:                  80685
Total generated tokens (retokenized):    80682
Request throughput (req/s):              0.43
Input token throughput (tok/s):          261.14
Output token throughput (tok/s):         135.37
Total token throughput (tok/s):          396.51
Concurrency:                             15.52
----------------End-to-End Latency----------------
Mean E2E Latency (ms):                   36130.29
Median E2E Latency (ms):                 35254.99
---------------Time to First Token----------------
Mean TTFT (ms):                          1411.06
Median TTFT (ms):                        1085.16
P99 TTFT (ms):                           6593.62
---------------Inter-Token Latency----------------
Mean ITL (ms):                           110.52
Median ITL (ms):                         69.21
P95 ITL (ms):                            217.64
P99 ITL (ms):                            1289.18
Max ITL (ms):                            6730.86
==================================================
```

### Comparison
  | Metric           | Flag Enabled | Baseline     | Difference |
  |------------------|--------------|--------------|------------|
  | Throughput       | 387.51 tok/s | 396.51 tok/s | -2.3%      |
  | Request/s        | 0.42         | 0.43         | -2.3%      |
  | Mean E2E Latency | 36,988 ms    | 36,130 ms    | +2.4%      |
  | Mean TTFT        | 1,459 ms     | 1,411 ms     | +3.4%      |
  | Mean ITL         | 113.09 ms    | 110.52 ms    | +2.3%      |

With random dataset and variance, we see there is no statistically significant performance impact. I have run several other benchmarks also showing ~+-2% differences, which can go either way depending on the randomization and what else is running on my system at the time.

<!-- If this pull request impacts inference speed, provide benchmarking and profiling results. -->

## Checklist

- [x] Format your code according to the [Format code with pre-commit](https://docs.sglang.ai/developer_guide/contribution_guide.html#format-code-with-pre-commit).
- [x] Add unit tests according to the [Run and add unit tests](https://docs.sglang.ai/developer_guide/contribution_guide.html#run-and-add-unit-tests).
- [x] Update documentation according to [Write documentations](https://docs.sglang.ai/developer_guide/contribution_guide.html#write-documentations).
- [x] Provide accuracy and speed benchmark results according to [Test the accuracy](https://docs.sglang.ai/developer_guide/contribution_guide.html#test-the-accuracy) and [Benchmark the speed](https://docs.sglang.ai/developer_guide/contribution_guide.html#benchmark-the-speed).