@gholmes829 2025-12-12
@XucSh sure thing.

I've re-applied my changes on top of latest main and pushed over original commits, since the conflicts with new mixin pattern were pretty significant.

From what I can tell, my changes seem to be compatible with the new async and PD disaggregation changes, but please call out if not the case.

Also fixed a gap in the original where PP sleep-on-idle only worked with DP attention - now works for regular DP+PP as well.

Haven't had time to test locally just yet, so lets see how the pipeline tests go. Looks like they are rate limited, so I will trigger them again in a bit.

Looking forward to the review!